refactor: api depend on core not internal
The api package depends on the internal/storage package. This makes
it very difficult for consumers of the package to be able to work
with juju. The solution is to move the parsing of the storage
directives to the core package. Everything can stay internal.
