+++ api/client/application/client_test.go
@@ -22,8 +22,8 @@ import (
 	"github.com/juju/juju/core/crossmodel"
 	"github.com/juju/juju/core/instance"
 	"github.com/juju/juju/core/relation"
+	"github.com/juju/juju/core/storage"
 	"github.com/juju/juju/internal/charm"
-	"github.com/juju/juju/internal/storage"
 	coretesting "github.com/juju/juju/internal/testing"
 	jujutesting "github.com/juju/juju/juju/testing"
 	"github.com/juju/juju/rpc/params"
