+++ cmd/juju/application/deployer/charm.go
@@ -23,9 +23,10 @@ import (
 	"github.com/juju/juju/core/devices"
 	"github.com/juju/juju/core/instance"
 	"github.com/juju/juju/core/model"
+	"github.com/juju/juju/core/storage"
 	"github.com/juju/juju/internal/charm"
 	"github.com/juju/juju/internal/cmd"
-	"github.com/juju/juju/internal/storage"
+	internalstorage "github.com/juju/juju/internal/storage"
 	"github.com/juju/juju/internal/storage/provider"
 )
 
@@ -89,7 +90,7 @@ func (d *deployCharm) deploy(
 			continue
 		}
 		for _, s := range d.storage {
-			if !provider.AllowedContainerProvider(storage.ProviderType(s.Pool)) {
+			if !provider.AllowedContainerProvider(internalstorage.ProviderType(s.Pool)) {
 				return errors.NotSupportedf("adding storage of type %q to %s container", s.Pool, string(t))
 			}
 		}
