+++ cmd/juju/application/deployer/deployer.go
@@ -29,11 +29,11 @@ import (
 	"github.com/juju/juju/core/devices"
 	"github.com/juju/juju/core/instance"
 	"github.com/juju/juju/core/model"
+	"github.com/juju/juju/core/storage"
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/internal/charm"
 	charmresource "github.com/juju/juju/internal/charm/resource"
 	internallogger "github.com/juju/juju/internal/logger"
-	"github.com/juju/juju/internal/storage"
 )
 
 var logger = internallogger.GetLogger("juju.cmd.juju.application.deployer")
