+++ cmd/juju/application/flags.go
@@ -11,7 +11,7 @@ import (
 	"github.com/juju/names/v6"
 
 	"github.com/juju/juju/core/devices"
-	"github.com/juju/juju/internal/storage"
+	"github.com/juju/juju/core/storage"
 )
 
 type storageFlag struct {
