+++ core/storage/directive_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/juju/tc"
 
-	"github.com/juju/juju/internal/storage"
+	"github.com/juju/juju/core/storage"
 	"github.com/juju/juju/internal/testing"
 )
