+++ rpc/params/applications.go
@@ -9,7 +9,7 @@ import (
 	"github.com/juju/juju/core/constraints"
 	"github.com/juju/juju/core/devices"
 	"github.com/juju/juju/core/instance"
-	"github.com/juju/juju/internal/storage"
+	"github.com/juju/juju/core/storage"
 )
 
 // ApplicationsDeploy holds the parameters for deploying one or more applications.
